+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/integer/BitShiftNode.java
@@ -27,7 +27,7 @@ public abstract class BitShiftNode extends IntegerNode {
   private final CountingConditionProfile rightShiftExceedsLongWidth =
       CountingConditionProfile.create();
 
-  abstract Object execute(Object self, Object that);
+  abstract Object execute(Object own, Object that);
 
   @NeverDefault
   static BitShiftNode build() {